+++ annotations/nocturne.exe/pseudocode/src/core/charactr.cpp/CCharacter_canLookAt_FUN_00429fe0.cpp
@@ -24,8 +24,8 @@ uint core_charactr_cpp_CCharacter_canLookAt_FUN_00429fe0(int param_1)
                            ((CMotionController *)(param_1 + 0x150));
       iVar1 = core_motion_cpp_CMotionList_findStateIndex_FUN_004e10a0(this_ptr,state_name,iVar1);
       if (-1 < iVar1) {
-        fVar3 = (float)core_motion_cpp_CMotionController_getStateBlendWeight_FUN_004e18d0
-                                 ((CMotionController *)(param_1 + 0x150),iVar1);
+        fVar3 = core_motion_cpp_CMotionController_getStateBlendWeight_FUN_004e18d0
+                          ((CMotionController *)(param_1 + 0x150),iVar1);
         if ((float)0.94999999999999996 < fVar3) {
           return 0;
         }
